more fixes based on pull request comments
